refactor: colocate testing utils
