+++ src/formats/speedscope/index.test.ts
@@ -1,16 +1,14 @@
 import { describe, expect, test } from 'vitest'
-import { defaultShowEntry, normalizeProfileToMdOptions } from '../../options.ts'
 import {
   callersTables,
-  categoryTables,
   linesTables,
-  profileTitles,
-  selfSizeTables,
   selfTimeTables,
-  summaryLines,
   totalTimeTables,
-} from '../../testing/markdown.ts'
-import { convertJsonToMd } from '../testing/convert.ts'
+} from '../../modalities/profile/testing.ts'
+import { selfSizeTables } from '../../modalities/snapshot/testing.ts'
+import { defaultShowEntry, normalizeProfileToMdOptions } from '../../options.ts'
+import { categoryTables, profileTitles, summaryLines } from '../../testing.ts'
+import { convertJsonToMd } from '../testing.ts'
 import { speedscopeConverter } from './index.ts'
 import {
   makeEventedProfile,
