+++ src/formats/v8/cpu-profile/index.test.ts
@@ -1,16 +1,16 @@
 import { describe, expect, test } from 'vitest'
 import {
-  defaultShowEntry,
-  normalizeProfileToMdOptions,
-} from '../../../options.ts'
-import {
   callersTables,
-  categoryTables,
   linesTables,
   selfTimeTables,
   totalTimeTables,
-} from '../../../testing/markdown.ts'
-import { convertJsonToMd } from '../../testing/convert.ts'
+} from '../../../modalities/profile/testing.ts'
+import {
+  defaultShowEntry,
+  normalizeProfileToMdOptions,
+} from '../../../options.ts'
+import { categoryTables } from '../../../testing.ts'
+import { convertJsonToMd } from '../../testing.ts'
 import { v8CpuProfileConverter } from './index.ts'
 import { makeV8CpuProfileRoot } from './testing.ts'
