+++ src/source-map.test.ts
@@ -5,7 +5,7 @@ import type {
 } from './options.ts'
 import { normalizeSourceMaps, sourceMapSourceLocation } from './source-map.ts'
 import type { SourceMap } from './source-map.ts'
-import { resolveProfileToMdOptions } from './testing/options.ts'
+import { resolveProfileToMdOptions } from './testing.ts'
 
 // Maps generated line 1 col 0 -> sources[0] line 1 col 0 (0-based).
 const L1_C0_TO_SOURCE_0_L1_C0 = `AAAA`
